+++ packages/typespec-client-generator-core/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./context.js";
 export * from "./decorators.js";
-export { EXACT_NAME_PREFIX, hasExactNameMarker, normalizeExactName } from "./functions.js";
+export { hasExactNameMarker, normalizeExactName } from "./functions.js";
 export * from "./interfaces.js";
 export * from "./lib.js";
 export { $linter } from "./linter.js";