+++ packages/typespec-client-generator-core/src/linter.ts
@@ -1,11 +1,17 @@
 import { defineLinter } from "@typespec/compiler";
+import { csharpNoUrlSuffixRule } from "./rules/csharp-no-url-suffix.js";
 import { noUnnamedTypesRule } from "./rules/no-unnamed-types.rule.js";
 import { propertyNameConflictRule } from "./rules/property-name-conflict.rule.js";
 import { requireClientSuffixRule } from "./rules/require-client-suffix.rule.js";
 
-const rules = [requireClientSuffixRule, propertyNameConflictRule, noUnnamedTypesRule];
+const rules = [
+  requireClientSuffixRule,
+  propertyNameConflictRule,
+  noUnnamedTypesRule,
+  csharpNoUrlSuffixRule,
+];
 
-const csharpRules = [propertyNameConflictRule];
+const csharpRules = [propertyNameConflictRule, csharpNoUrlSuffixRule];
 
 export const $linter = defineLinter({
   rules,