+++ packages/typespec-ts/src/meta-tree.ts
@@ -1,0 +1,8 @@
+import { Type } from "@typespec/compiler";
+import { Schema } from "./interfaces.js";
+
+export interface ClientTypeMetadata {
+  clientType: Schema;
+}
+
+export type ClientTypeMetaTree = Map<Type, ClientTypeMetadata>;